ai-services/requirements.txt
# AI/ML Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP & API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Redis & Caching
redis==5.0.1
celery==5.3.4

# Machine Learning & AI
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.4

# NLP & Language Models
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
tokenizers==0.15.0
huggingface-hub==0.19.4

# OpenAI Integration
openai==1.3.7

# Data Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Recommendation Systems
implicit==0.7.2
surprise==1.1.3

# Geospatial & Location
geopy==2.4.1
folium==0.15.0

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3
pycountry==22.3.13

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Image Processing (for travel photos)
Pillow==10.1.0
opencv-python==4.8.1.78

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API Documentation
fastapi-users==12.1.2
fastapi-limiter==0.1.5

# Firebase Admin SDK
firebase-admin==6.2.0

# OpenAI API
openai==1.3.7
